refactor

+++ database/migrations/0000_00_00_000001_create_jba_profile_regions_table.php
@@ -13,16 +13,21 @@ class CreateJbaProfileRegionsTable extends Migration
      */
     public function up()
     {
-        Schema::create($this->getTableName('regions'), function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('parent_id')->unsigned()->nullable();
-            $table->string('name', 32);
+        if(!Schema::hasTable($this->getTableName('regions'))) {
+            Schema::create($this->getTableName('regions'), function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->bigInteger('parent_id')->unsigned()->nullable();
+                $table->string('name', 32);
 
-            $table->foreign('parent_id')
-                ->references('id')
-                ->on($this->getTableName('regions'))
-                ->onDelete('cascade');
-        });
+                $table->foreign('parent_id')
+                    ->references('id')
+                    ->on($this->getTableName('regions'))
+                    ->onDelete('cascade');
+            });
+        }
+        else {
+            dump('Ignore' . __FILE__ . ' as ' . $this->getTableName('regions') . ' table has exists');
+        }
     }
 
     /**
