+++ database/migrations/0000_00_00_000003_create_jba_profile_profiles_table.php
@@ -13,33 +13,39 @@ class CreateJbaProfileProfilesTable extends Migration
      */
     public function up()
     {
-        Schema::create($this->getTableName('profiles'), function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('user_id')->unsigned();
-            $table->string('first_name', 64)->nullable();
-            $table->string('last_name', 64)->nullable();
-            $table->string('phone', 32)->nullable();
-            $table->string('gender', 16)->nullable();
-            $table->string('avatar', 128)->nullable();
-            //    $table->string('msd_number')->nullable();
-            $table->string('location', 256)->nullable();
-            $table->bigInteger('region_id')->unsigned()->nullable();
-            $table->date('date_of_birth')->nullable();
-            $table->string('ethnicity', 32)->nullable();
-            $table->string('residency', 32)->nullable();
-            $table->boolean('employment_status')->nullable();
-            $table->timestamps();
+        if(!Schema::hasTable($this->getTableName('profiles'))) {
 
-            /*$table->foreign('user_id')
-                ->references('id')
-                ->on($this->getTableName('users'))
-                ->onDelete('cascade');*/
+            Schema::create($this->getTableName('profiles'), function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->bigInteger('user_id')->unsigned();
+                $table->string('first_name', 64)->nullable();
+                $table->string('last_name', 64)->nullable();
+                $table->string('phone', 32)->nullable();
+                $table->string('gender', 16)->nullable();
+                $table->string('avatar', 128)->nullable();
+                //    $table->string('msd_number')->nullable();
+                $table->string('location', 256)->nullable();
+                $table->bigInteger('region_id')->unsigned()->nullable();
+                $table->date('date_of_birth')->nullable();
+                $table->string('ethnicity', 32)->nullable();
+                $table->string('residency', 32)->nullable();
+                $table->boolean('employment_status')->nullable();
+                $table->timestamps();
 
-            $table->foreign('region_id')
-                ->references('id')
-                ->on($this->getTableName('regions'))
-                ->onDelete('cascade');
-        });
+                /*$table->foreign('user_id')
+                    ->references('id')
+                    ->on($this->getTableName('users'))
+                    ->onDelete('cascade');*/
+
+                $table->foreign('region_id')
+                    ->references('id')
+                    ->on($this->getTableName('regions'))
+                    ->onDelete('cascade');
+            });
+        }
+        else {
+            dump('Ignore' . __FILE__ . ' as ' . $this->getTableName('profiles') . ' table has exists');
+        }
     }
 
     /**
